fix: 로직분리
fix: 로직분리
