+++ becok/public/index.html
@@ -4,6 +4,7 @@
   <head>
     <meta charset="UTF-8" />
     <title>becok</title>
+    <link rel="icon" type="image/png" href="%PUBLIC_URL%/becok.png" />
   </head>
 
   <body>
